requirements.txt
requests
python-dotenv
fastapi
uvicorn
pydantic
starlette
diagrams
fastapi-mcp
pytest
pytest-cov
pytest-asyncio
pytest-mock
pytest-httpx
pytest-mypy
trio

protobuf
tiktoken
boto3
google-genai
ollama
anthropic
openai
toml
Jinja2
groq
langchain
langchain-community
langchain-chroma
google-adk
